+++ src/Builtins/String.cpp
@@ -68,7 +68,7 @@ namespace Ark::internal::Builtins::String
             throw Ark::TypeError(STR_RM_TE1);
 
         long id = static_cast<long>(n[1].number());
-        if (id < 0 || id >= n[0].string_ref().size())
+        if (id < 0 || id >= static_cast<long>(n[0].string_ref().size()))
             throw std::runtime_error(STR_RM_OOR);
 
         n[0].string_ref().erase(id, id + 1);
